Added open Ai method and changed controller
